cambios

+++ src/resource/catalogos/catalogo_categorias/catalogo_categorias.service.ts
@@ -6,6 +6,11 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CatalogoCategoria } from './entities/catalogo_categoria.entity';
 
+import { User_Interface } from 'src/common/interfaces/user.interface';
+import { validateOwnershipAdmin } from 'src/Guard/validateOwnerShip.guard';
+import { Errores_Catalogos } from 'src/common/helpers/Errores.service';
+import { Exito_Catalogos } from 'src/common/helpers/Confirmaciones.service';
+
 @Injectable()
 export class CatalogoCategoriasService {
 
@@ -14,23 +19,55 @@ export class CatalogoCategoriasService {
     private catalogoCategoriaRepository: Repository<CatalogoCategoria>
   ) {}
 
-  create(createCatalogoCategoriaDto: CreateCatalogoCategoriaDto) {
-    return this.catalogoCategoriaRepository.save(createCatalogoCategoriaDto);
+  create(createCatalogoCategoriaDto: CreateCatalogoCategoriaDto, user: User_Interface) {
+    validateOwnershipAdmin(user);
+
+    let buscar = this.catalogoCategoriaRepository.findOne({
+      where: { categoria: createCatalogoCategoriaDto.categoria },
+    });
+
+    if (buscar) {
+      return Errores_Catalogos.CATALOG_ALREADY_EXISTS;
+    } else {
+      this.catalogoCategoriaRepository.save(createCatalogoCategoriaDto);
+      return Exito_Catalogos.CATALOGO_CREADO;
+    }
   }
 
-  findAll() {
+  findAll( user: User_Interface) {
+    validateOwnershipAdmin(user);
     return this.catalogoCategoriaRepository.find();
   }
 
-  findOne(id: number) {
-    return this.catalogoCategoriaRepository.findOneById(id);
+  findOne(id: number, user: User_Interface) {
+    validateOwnershipAdmin(user);
+
+    try {
+      return this.catalogoCategoriaRepository.findOneById(id);
+    } catch (error) {
+      return Errores_Catalogos.CATALOG_NOT_FOUND;
+    }
   }
 
-  update(id: number, updateCatalogoCategoriaDto: UpdateCatalogoCategoriaDto) {
-    return this.catalogoCategoriaRepository.update(id, updateCatalogoCategoriaDto);
+  update(id: number, updateCatalogoCategoriaDto: UpdateCatalogoCategoriaDto, user: User_Interface) {
+    validateOwnershipAdmin(user);
+
+    try {
+      this.catalogoCategoriaRepository.update(id, updateCatalogoCategoriaDto);
+      return Exito_Catalogos.CATALOGO_ACTUALIZADO;
+    } catch (error) {
+      return Errores_Catalogos.CATALOG_NOT_FOUND;
+    }
   }
 
-  remove(id: number) {
-    return this.catalogoCategoriaRepository.delete(id);
+  remove(id: number, user: User_Interface) {
+    validateOwnershipAdmin(user);
+
+    try {
+      this.catalogoCategoriaRepository.delete(id);
+      return Exito_Catalogos.CATALOGO_ELIMINADO;
+    } catch (error) {
+      return Errores_Catalogos.CATALOG_NOT_FOUND;
+    }
   }
 }
